requirements.txt
curl_cffi>=0.5.0
pywidevine>=1.8.0
pycryptodome>=3.19.0
protobuf>=4.0.0
